fix: 할당되지 않은 프로젝트 조회 가능하도록 수정
본인에게 할당 된 프로젝트만 조회 체크박스 추가가

+++ src/components/project/details/ProjectHeader.tsx
@@ -37,9 +37,11 @@ export default function ProjectHeader() {
   }
 
   const handleDeleteProject = async () => {
-    await deleteProject(Number(id))
-    queryClient.invalidateQueries({ queryKey: ['project', 'list'] })
-    router.push('/dashboard/projects')
+    if(confirm('정말 삭제하시겠습니까?')) {
+      await deleteProject(Number(id))
+      queryClient.invalidateQueries({ queryKey: ['project', 'list'] })
+      router.push('/dashboard/projects')
+    }
   }
 
   return (
